Parameter handling

+++ src/routes.ts
@@ -1,4 +1,3 @@
-// import cors from '@koa/cors';
 import Router from '@koa/router';
 import koaBody from 'koa-body';
 import fs from 'node:fs';
@@ -9,7 +8,7 @@ import { getFileFormat } from './filetype.js';
 import { log } from './log.js';
 import { createResizedImages, resizedKeys } from './resizing.js';
 import { activeStorage, incomingStorage, tempStorage } from './storage.js';
-import { createUniqueKey, multiParams, uniqueParam } from './utils.js';
+import { generateUniqueKeyPrefix, isFileParameterValid, isKeyParameterValid, isKeysParameterValid } from './utils.js';
 
 export const router = new Router();
 
@@ -21,21 +20,20 @@ router.get('/ping', async ctx => {
 // Uploaded images are verified, resized images generated and stored in the incoming storage
 // This endpoint is called directly by the UI.
 router.post('/upload', koaBody({ multipart: true }), async ctx => {
-  const keyPrefix = createUniqueKey();
-  const file = ctx.request.files && ctx.request.files['file'];
+  const keyPrefix = generateUniqueKeyPrefix();
+  const file = ctx.request.files?.['file'];
 
-  if (Array.isArray(file)) {
-    return ctx.throw(400, `Bad request. Only one file allowed.`);
-  }
-  if (!file || file.size === 0) {
-    return ctx.throw(400, `Bad request. No file provided.`);
+  if (!isFileParameterValid(file)) {
+    ctx.throw(400, `Bad request. Exactly one file has to be provided.`);
+    return;
   }
 
   let format: string;
   try {
     format = getFileFormat(file.path);
   } catch (error: unknown) {
-    return ctx.throw(400, error instanceof Error ? error.message : `${error}`);
+    ctx.throw(400, error instanceof Error ? error.message : `${error}`);
+    return;
   }
 
   // rename with 'official' extension
@@ -61,20 +59,24 @@ router.post('/upload', koaBody({ multipart: true }), async ctx => {
   ctx.body = {
     filename: `${keyPrefix}.${format}`
   };
-  return;
 });
 
 // Publish images from incoming storage to active storage
 // Only called by api service
 router.post('/publish', koaBody({ multipart: true }), apiOnly, async ctx => {
   const { filename: key } = ctx.request.body;
-  uniqueParam(ctx, key, 'filename');
+
+  if (!isKeyParameterValid(key)) {
+    ctx.throw(400, 'Bad parameter "filename".');
+    return;
+  }
 
   const published = await activeStorage.exists(key);
   const incoming = await incomingStorage.exists(key);
 
   if (!published && !incoming) {
     ctx.throw(400, 'Unknown image.');
+    return;
   }
 
   if (!published) {
@@ -93,8 +95,14 @@ router.post('/publish', koaBody({ multipart: true }), apiOnly, async ctx => {
 // Delete a file from active storage
 // Only called by api service
 router.post('/delete', koaBody({ multipart: true }), apiOnly, async ctx => {
-  const { filenames: keys } = ctx.request.body;
-  for (const key of multiParams(ctx, keys, 'filenames')) {
+  let { filenames: keys } = ctx.request.body;
+  keys = Array.isArray(keys) ? keys : [keys];
+  if (!isKeysParameterValid(keys)) {
+    ctx.throw(400, 'Bad parameter "filenames"');
+    return;
+  }
+
+  for (const key of keys) {
     try {
       await activeStorage.delete(key);
     } catch {
